contentagent/requirements.txt
# MerchAgent-specific requirements.txt
# Only includes packages needed by MerchAgent

# Core Agent Engine dependencies
google-cloud-aiplatform[adk,agent_engines]>=1.96.0
google-cloud-secret-manager>=2.16.0
cloudpickle

# ADK and AI dependencies  
google-adk>=0.1.0
google-genai>=1.19.0
vertexai>=1.0.0

#Other specific dependencies
python-dotenv>=1.1.1
requests>=2.32.4
typing-extensions

# Optional (remove if not needed)
urllib3>=2.5.0

# Content Agent Specifics
moviepy==2.2.1
google-cloud>=0.34.0
google-cloud-secret-manager>=2.24.0
google-cloud-speech>=2.33.0
google-cloud-storage>=2.19.0
google-cloud-texttospeech>=2.27.0